+++ hub/router/handler.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"GADS/hub/auth"
-	"GADS/hub/hubutil"
+	"GADS/hub/devices"
 	"github.com/gin-contrib/cors"
 	"github.com/gin-contrib/static"
 	"github.com/gin-gonic/gin"
@@ -18,11 +18,11 @@ func HandleRequests(authentication bool) *gin.Engine {
 	config.AllowHeaders = []string{"X-Auth-Token", "Content-Type"}
 	r.Use(cors.New(config))
 
-	indexHtmlPath := filepath.Join(hubutil.ConfigData.UIFilesTempDir, "index.html")
+	indexHtmlPath := filepath.Join(devices.ConfigData.UIFilesTempDir, "index.html")
 
 	// Configuration for SAP applications
 	// Serve the static files from the built React app
-	r.Use(static.Serve("/", static.LocalFile(hubutil.ConfigData.UIFilesTempDir, true)))
+	r.Use(static.Serve("/", static.LocalFile(devices.ConfigData.UIFilesTempDir, true)))
 	// For any missing route serve the index.htm from the static files
 	// This will fix the issue with accessing particular endpoint in the browser manually or with refresh
 	r.NoRoute(func(c *gin.Context) {
